// src/types/generated/SchedulerEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AttentionId } from "./AttentionId";
import type { BreakId } from "./BreakId";

/**
 * Events sent to the frontend for display
 */
export type SchedulerEvent = { "type": "miniBreak", "data": BreakId } | { "type": "longBreak", "data": BreakId } | { "type": "attention", "data": AttentionId };